requirements.txt
streamlit
python-dotenv
langchain
googletrans==4.0.0-rc1
pytube
faiss-cpu
openai
tiktoken